Sync the repo with the ML-Agents

+++ com.unity.ml-agents/Runtime/Sensors/Reflection/IntReflectionSensor.cs
@@ -8,12 +8,12 @@ namespace Unity.MLAgents.Sensors.Reflection
     {
         public IntReflectionSensor(ReflectionSensorInfo reflectionSensorInfo)
             : base(reflectionSensorInfo, 1)
-        {}
+        { }
 
         internal override void WriteReflectedField(ObservationWriter writer)
         {
             var intVal = (System.Int32)GetReflectedValue();
-            writer[0] = (float)intVal;
+            writer[0] = intVal;
         }
     }
 }
